+++ templates/movie_status_popup.py
@@ -3,7 +3,7 @@ import json
 import core
 import dominate
 from core import sqldb
-from core.conversions import Conversions
+from core.helpers import Conversions
 from dominate.tags import *
